backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.6.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
stockfish==3.28.0
supabase==1.2.0
email-validator==2.1.0.post1
httpx==0.24.1
gotrue==1.2.0
pinecone
openai==1.12.0
python-chess==1.999.0